v2/requirements.txt
fastapi
uvicorn
google-cloud-storage
toolz